refactor: extract parseCollegeQuery and support short ranks
