+++ src/lib/parseCollegeQuery.js
@@ -1,0 +1,33 @@
+export const categoryMap = {
+  obc: 'OBC-NCL',
+  'obc ncl': 'OBC-NCL',
+  sc: 'SC',
+  st: 'ST',
+  ews: 'EWS',
+  gen: 'OPEN',
+  general: 'OPEN',
+  open: 'OPEN'
+};
+
+export function parseCollegeQuery(text) {
+  const lower = text.toLowerCase();
+  let match =
+    lower.match(/\brank(?:\s*is)?\s*#?(\d{1,4})\b/) ||
+    lower.match(/#?(\d{1,4})\s*rank\b/);
+  if (!match) {
+    match = lower.match(/\b(\d{3,})\b/);
+  }
+  const rankStr = match && (match[1] || match[2]);
+  const rank = rankStr ? parseInt(rankStr, 10) : null;
+
+  let category = null;
+  for (const key of Object.keys(categoryMap)) {
+    if (lower.includes(key)) { category = categoryMap[key]; break; }
+  }
+  const branchMatch = text.match(/\b(CSE|Computer Science|ECE|Electrical|Electronics|Mechanical|Civil|IT|Information Technology)\b/i);
+  const branch = branchMatch ? branchMatch[0] : null;
+  const instituteMatch = text.match(/(?:at|in|for)\s+([A-Za-z ]*(?:IIT|NIT|IIIT)[A-Za-z ]*)/i);
+  const institute = instituteMatch ? instituteMatch[1].trim() : null;
+  const isCollegeQuery = rank !== null || branch || institute || lower.includes('college');
+  return { rank, category, branch, institute, isCollegeQuery };
+}
